move LoadState to core to multiple use
