+++ feature_recipes/src/main/java/com/example/shoplist/feature_recipes/ui/RecipeSearchFragment.kt
@@ -5,13 +5,13 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.GridLayoutManager
 import by.kirich1409.viewbindingdelegate.viewBinding
+import com.example.shoplist.core.models.LoadState
 import com.example.shoplist.core.ui.showErrorMessage
 import com.example.shoplist.core.ui.setVisibility
 import com.example.shoplist.domain.models.MealsEntity
 import com.example.shoplist.feature_meal_item.ui.MealsAdapter
 import com.example.shoplist.feature_recipes.R
 import com.example.shoplist.feature_recipes.databinding.FragmentRecipeSearchBinding
-import com.example.shoplist.feature_recipes.models.LoadState
 import com.example.shoplist.feature_recipes.viewModel.RecipeSearchViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
