+++ feature_shoplist/src/main/java/com/example/shoplist/feature_shoplist/viewModel/ShoplistViewModel.kt
@@ -3,10 +3,10 @@ package com.example.shoplist.feature_shoplist.viewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.shoplist.core.models.LoadState
 import com.example.shoplist.domain.models.Errors
 import com.example.shoplist.domain.models.ShoplistEntity
 import com.example.shoplist.domain.repos.LoadingShoplistRepo
-import com.example.shoplist.feature_shoplist.models.LoadState
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
